python_ncbi_dbapi post_build.py: Conditionalize cleanup of .../lib/python_ncbi_dbapi.so, since os.unlink throws an exception when its argument doesn't exist (as on Windows).
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@60169 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/projects/python_ncbi_dbapi/post_build/post_build.py
@@ -7,7 +7,8 @@ import glob, os, sys
 # needing to match both Unix and Windows library filenames.
 for x in glob.glob(sys.argv[1] + '/lib/*xxconnect*'):
     os.unlink(x)
-os.unlink(sys.argv[1] + '/lib/python_ncbi_dbapi.so')
+if os.path.exists(sys.argv[1] + '/lib/python_ncbi_dbapi.so'):
+    os.unlink(sys.argv[1] + '/lib/python_ncbi_dbapi.so')
 
 # Build a tarball for system-wide installation only on Linux.
 if sys.platform == 'linux2':
